Update petopia (4).sql
The database now has a table that stores different messages that users send. This table is called contactforms. You do not need to have an account with us to send a message this way, just your name and email.

I have also updated the review table. I have included product_ID so that each review can be correctly mapped onto each product. I have also included a rating column, so that users can rate the product /5. You will need to have an account with us to send a review, as customer id is required.